--- packages/ui/react/src/styles/props/flex.props.ts
@@ -1,21 +1,0 @@
-import { type StyleProps, createRawStyleProps } from '@favolink-ui/system';
-
-export type FlexStyleProps = StyleProps<
-  | 'alignItems'
-  | 'flexBasis'
-  | 'flexDirection'
-  | 'flexGrow'
-  | 'flexShrink'
-  | 'flexWrap'
-  | 'justifyContent'
->;
-
-export const flexStyleProps = createRawStyleProps<FlexStyleProps>(
-  'alignItems',
-  'flexBasis',
-  'flexDirection',
-  'flexGrow',
-  'flexShrink',
-  'flexWrap',
-  'justifyContent',
-);
