--- packages/ui/react/src/styles/props/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export { flexStyleProps, type FlexStyleProps } from './flex.props';
