--- packages/ui/react/src/styles/utils/dynamic/index.ts
@@ -1,86 +1,0 @@
-import { flexBasis, flexBasisVar } from './flex-basis.css';
-import { gap, gapVar, gapX, gapXVar, gapY, gapYVar } from './gap.css';
-import { height, heightVar } from './height.css';
-import {
-  bottom,
-  bottomVar,
-  inset,
-  insetVar,
-  left,
-  leftVar,
-  right,
-  rightVar,
-  top,
-  topVar,
-} from './inset.css';
-import { maxHeight, maxHeightVar } from './max-height.css';
-import { maxWidth, maxWidthVar } from './max-width.css';
-import { minHeight, minHeightVar } from './min-height.css';
-import { minWidth, minWidthVar } from './min-width.css';
-import {
-  padding,
-  paddingBottom,
-  paddingBottomVar,
-  paddingLeft,
-  paddingLeftVar,
-  paddingRight,
-  paddingRightVar,
-  paddingTop,
-  paddingTopVar,
-  paddingVar,
-  paddingX,
-  paddingXVar,
-  paddingY,
-  paddingYVar,
-} from './padding.css';
-import { width, widthVar } from './width.css';
-
-export const dynamicVars = {
-  flexBasisVar,
-  gapVar,
-  gapXVar,
-  gapYVar,
-  heightVar,
-  insetVar,
-  topVar,
-  bottomVar,
-  leftVar,
-  rightVar,
-  maxHeightVar,
-  maxWidthVar,
-  minHeightVar,
-  minWidthVar,
-  paddingVar,
-  paddingTopVar,
-  paddingBottomVar,
-  paddingRightVar,
-  paddingLeftVar,
-  paddingXVar,
-  paddingYVar,
-  widthVar,
-};
-
-export const dynamicStyles = {
-  flexBasis,
-  gap,
-  gapX,
-  gapY,
-  height,
-  inset,
-  top,
-  bottom,
-  left,
-  right,
-  maxHeight,
-  maxWidth,
-  minHeight,
-  minWidth,
-  padding,
-  paddingTop,
-  paddingBottom,
-  paddingRight,
-  paddingLeft,
-  paddingX,
-  paddingY,
-  width,
-};
