--- packages/ui/react/src/styles/utils/enum/index.ts
@@ -1,37 +1,0 @@
-import { alignItems } from './align-items.css';
-import { body } from './body.css';
-import { borderRadius } from './border-radius.css';
-import { display } from './display.css';
-import { flexDirection } from './flex-direction.css';
-import { flexGrow } from './flex-grow.css';
-import { flexShrink } from './flex-shrink.css';
-import { flexWrap } from './flex-wrap.css';
-import { heading } from './heading.css';
-import { justifyContent } from './justify-content.css';
-import { letterSpacing } from './letter-spacing.css';
-import { overflow, overflowX, overflowY } from './overflow.css';
-import { position } from './position.css';
-import { textAlign } from './text-align.css';
-import { textWrap } from './text-wrap.css';
-import { truncate } from './truncate.css';
-
-export const enumStyles = {
-  alignItems,
-  body,
-  borderRadius,
-  display,
-  flexDirection,
-  flexShrink,
-  flexGrow,
-  flexWrap,
-  heading,
-  justifyContent,
-  letterSpacing,
-  overflow,
-  overflowX,
-  overflowY,
-  position,
-  textAlign,
-  textWrap,
-  truncate,
-};
