--- packages/ui/react/src/styles/vars/flex.vars.css.ts
@@ -1,25 +1,0 @@
-import { createVar } from '@vanilla-extract/css';
-
-const alignItems = createVar();
-
-const flexBasis = createVar();
-
-const flexDirection = createVar();
-
-const flexGrow = createVar();
-
-const flexShrink = createVar();
-
-const flexWrap = createVar();
-
-const justifyContent = createVar();
-
-export const flexStyleVars = {
-  alignItems,
-  flexBasis,
-  flexDirection,
-  flexGrow,
-  flexShrink,
-  flexWrap,
-  justifyContent,
-};
