--- packages/ui/react/src/styles/vars/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export { flexStyleVars } from './flex.vars.css';
