+++ packages/ui/styles/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@favolink-ui/styles",
+  "version": "0.0.0",
+  "private": true,
+  "exports": {
+    ".": "./src/index.ts",
+    "./reset.css": "./src/reset.css.ts",
+    "./theme.css": "./src/theme.css.ts"
+  },
+  "scripts": {
+    "lint": "eslint \"**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}\""
+  },
+  "dependencies": {
+    "@favolink-ui/utils": "workspace:*",
+    "@vanilla-extract/css": "^1.14.1",
+    "@vanilla-extract/dynamic": "^2.1.0",
+    "react": "^18.2.0"
+  },
+  "devDependencies": {
+    "@types/react": "^18.3.3"
+  }
+}
