+++ packages/ui/styles/src/index.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export * from './props';
-export * from './vars';
+export * from './utils';
 export * from './extract-dynamic-props';
 export * from './make-style-variants-data';
 export * from './theme.css';
