+++ packages/ui/styles/src/utils/enum/flex-shrink.css.ts
