+++ packages/ui/styles/src/utils/enum/text-align.css.ts
