+++ packages/ui/styles/src/utils/enum/truncate.css.ts
