+++ packages/ui/styles/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@tsconfig/favolink/base.json",
+  "include": [".eslintrc.js", "src"]
+}
